# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Umrah Travel ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [accounting]
      summary: List invoices
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: clientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [accounting]
      summary: Create invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [accounting]
      summary: Get invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [accounting]
      summary: Update invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}/payments:
    post:
      operationId: recordPayment
      tags: [accounting]
      summary: Record payment on invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /expenses:
    get:
      operationId: listExpenses
      tags: [accounting]
      summary: List expenses
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [accounting]
      summary: Create expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"

  # ── DASHBOARD ─────────────────────────────────────────────────────
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List generated documents
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
  /documents/generate:
    post:
      operationId: generateDocument
      tags: [documents]
      summary: Generate a document (voucher, invoice, receipt)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentGenerateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # ── USERS ──────────────────────────────────────────────────────
    User:
      type: object
      required: [id, name, email, role, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, sales, visa, accounts, support]
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
...
    Invoice:
      type: object
      required: [id, invoiceNumber, type, clientId, clientName, amount, currency, status, dueDate, createdAt]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        type:
          type: string
          enum: [customer, vendor]
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        vendorId:
          type: ["integer", "null"]
        vendorName:
          type: ["string", "null"]
        quotationId:
          type: ["integer", "null"]
        amount:
          type: number
        paidAmount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [draft, sent, partial, paid, overdue, cancelled]
        dueDate:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    InvoiceInput:
      type: object
      required: [type, amount, currency, dueDate]
      properties:
        type:
          type: string
          enum: [customer, vendor]
        clientId:
          type: integer
        vendorId:
          type: integer
        quotationId:
          type: integer
        amount:
          type: number
        currency:
          type: string
        dueDate:
          type: string
        notes:
          type: string
    InvoiceUpdate:
      type: object
      properties:
        status:
          type: string
        dueDate:
          type: string
        notes:
          type: string
    Payment:
      type: object
      required: [id, invoiceId, amount, currency, method, paidAt, createdAt]
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        amount:
          type: number
        currency:
          type: string
        method:
          type: string
          enum: [cash, bank_transfer, card, online]
        reference:
          type: ["string", "null"]
        paidAt:
          type: string
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [amount, currency, method, paidAt]
      properties:
        amount:
          type: number
        currency:
          type: string
        method:
          type: string
          enum: [cash, bank_transfer, card, online]
        reference:
          type: string
        paidAt:
          type: string
    Expense:
      type: object
      required: [id, title, category, amount, currency, date, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        amount:
          type: number
        currency:
          type: string
        vendorId:
          type: ["integer", "null"]
        vendorName:
          type: ["string", "null"]
        date:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required: [title, category, amount, currency, date]
      properties:
        title:
          type: string
        category:
          type: string
        amount:
          type: number
        currency:
          type: string
        vendorId:
          type: integer
        date:
          type: string
        notes:
          type: string

    # ── DASHBOARD ──────────────────────────────────────────────────
...
    Document:
      type: object
      required: [id, type, title, clientId, url, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [voucher, invoice, receipt, quotation_pdf]
        title:
          type: string
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        url:
          type: string
        createdAt:
          type: string
    DocumentGenerateInput:
      type: object
      required: [type, entityId]
      properties:
        type:
          type: string
          enum: [voucher, invoice, receipt, quotation_pdf]
        entityId:
          type: integer